Moved glue.qt.glue_application to glue.app.qt.application
