+++ glue/core/tests/test_state.py
@@ -8,9 +8,6 @@ import numpy as np
 
 from glue.external import six
 from glue import core
-from glue.qt.glue_application import GlueApplication
-from glue.viewers.histogram.qt import HistogramWidget
-from glue.viewers.scatter.qt import ScatterWidget
 from glue.tests.helpers import requires_astropy, make_file
 
 from ..data_factories import load_data
@@ -229,140 +226,6 @@ def test_polygonal_roi():
     r2 = clone(roi)
     assert r2.vx == [0, 0, 1]
     assert r2.vy == [0, 1, 0]
-
-
-def check_clone_app(app):
-    c = Cloner(app)
-    copy = c.us.object('__main__')
-
-    hub1 = app.session.hub
-    hub2 = copy.session.hub
-
-    assert len(hub1._subscriptions) == len(hub2._subscriptions)
-
-    # data collections are the same
-    for d1, d2 in zip(app.session.data_collection,
-                      copy.session.data_collection):
-        assert d1.label == d2.label
-        for cid1, cid2 in zip(d1.components, d2.components):
-            assert cid1.label == cid2.label
-
-            # order of components unspecified if label collisions
-            cid2 = c.get(cid1)
-            np.testing.assert_array_almost_equal(d1[cid1, 0:1],
-                                                 d2[cid2, 0:1], 3)
-
-    # same data viewers, in the same tabs
-    for tab1, tab2 in zip(app.viewers, copy.viewers):
-        assert len(tab1) == len(tab2)
-        for v1, v2 in zip(tab1, tab2):
-            assert type(v1) == type(v2)
-            # same window properties
-            assert v1.viewer_size == v2.viewer_size
-            assert v1.position == v2.position
-
-            # same viewer-level properties (axis label, scaling, etc)
-            assert set(v1.properties.keys()) == set(v2.properties.keys())
-            for k in v1.properties:
-                if hasattr(v1.properties[k], 'label'):
-                    assert v1.properties[k].label == v2.properties[k].label
-                else:
-                    assert v1.properties[k] == v2.properties[k] or \
-                        containers_equal(v1.properties[k], v2.properties[k])
-
-            assert len(v1.layers) == len(v2.layers)
-            for l1, l2 in zip(v1.layers, v2.layers):
-                assert l1.layer.label == l2.layer.label  # same data/subset
-                assert l1.layer.style == l2.layer.style
-
-    return copy
-
-
-class TestApplication(object):
-
-    def check_clone(self, app):
-        return check_clone_app(app)
-
-    def test_bare_application(self):
-        app = GlueApplication()
-        self.check_clone(app)
-
-    def test_data_application(self):
-        dc = core.DataCollection([core.Data(label='test',
-                                            x=[1, 2, 3], y=[2, 3, 4])])
-        app = GlueApplication(dc)
-        self.check_clone(app)
-
-    def test_links(self):
-        d1 = core.Data(label='x', x=[1, 2, 3])
-        d2 = core.Data(label='y', y=[3, 4, 8])
-        dc = core.DataCollection([d1, d2])
-        link = core.ComponentLink([d1.id['x']], d2.id['y'], doubler)
-        dc.add_link(link)
-
-        np.testing.assert_array_equal(d1['y'], [2, 4, 6])
-
-        app = GlueApplication(dc)
-        self.check_clone(app)
-
-    def test_scatter_viewer(self):
-        d = core.Data(label='x', x=[1, 2, 3, 4, 5], y=[2, 3, 4, 5, 6])
-        dc = core.DataCollection([d])
-        app = GlueApplication(dc)
-        w = app.new_data_viewer(ScatterWidget, data=d)
-        self.check_clone(app)
-
-        s1 = dc.new_subset_group()
-        s2 = dc.new_subset_group()
-        assert len(w.layers) == 3
-        l1, l2, l3 = w.layers
-        l1.zorder, l2.zorder = l2.zorder, l1.zorder
-        l3.visible = False
-        assert l3.visible is False
-        copy = self.check_clone(app)
-        assert copy.viewers[0][0].layers[-1].visible is False
-
-    def test_multi_tab(self):
-        d = core.Data(label='hist', x=[[1, 2], [2, 3]])
-        dc = core.DataCollection([d])
-
-        app = GlueApplication(dc)
-        w1 = app.new_data_viewer(HistogramWidget, data=d)
-        app.new_tab()
-        w2 = app.new_data_viewer(HistogramWidget, data=d)
-        assert app.viewers == ((w1,), (w2,))
-
-        self.check_clone(app)
-
-    def test_histogram(self):
-        d = core.Data(label='hist', x=[[1, 2], [2, 3]])
-        dc = core.DataCollection([d])
-
-        app = GlueApplication(dc)
-        w = app.new_data_viewer(HistogramWidget, data=d)
-        self.check_clone(app)
-
-        dc.new_subset_group()
-        assert len(w.layers) == 2
-        self.check_clone(app)
-
-        w.nbins = 7
-        self.check_clone(app)
-
-    def test_subset_groups_remain_synced_after_restore(self):
-        # regrssion test for 352
-        d = core.Data(label='hist', x=[[1, 2], [2, 3]])
-        dc = core.DataCollection([d])
-        dc.new_subset_group()
-        app = GlueApplication(dc)
-
-        app2 = clone(app)
-        sg = app2.data_collection.subset_groups[0]
-        assert sg.style.parent is sg
-
-        sg.style.color = '#112233'
-        assert sg.subsets[0].style.color == '#112233'
-
 
 class DummyClass(object):
     pass
