+++ glue/main.py
@@ -116,7 +116,7 @@ def die_on_error(msg):
 @die_on_error("Error restoring Glue session")
 def restore_session(gluefile):
     """Load a .glu file and return a DataCollection, Hub tuple"""
-    from glue.qt.glue_application import GlueApplication
+    from glue.app.qt import GlueApplication
     return GlueApplication.restore_session(gluefile)
 
 
@@ -150,7 +150,7 @@ def start_glue(gluefile=None, config=None, datafiles=None):
     :type datafiles: list of str
     """
     import glue
-    from glue.qt.glue_application import GlueApplication
+    from glue.app.qt import GlueApplication
 
     # Start off by loading plugins. We need to do this before restoring
     # the session or loading the configuration since these may use existing
