+++ glue/qt/widgets/layer_tree_widget.py
@@ -514,7 +514,7 @@ class LayerTreeWidget(QtGui.QWidget, Ui_LayerTree):
     def _load_data(self):
         """ Interactively loads data from a data set. Adds
         as new layer """
-        from glue.qt.glue_application import GlueApplication
+        from glue.app.qt import GlueApplication
 
         layers = qtutil.data_wizard()
         GlueApplication.add_datasets(self.data_collection, layers)
