+++ glue/tests/test_qglue.py
@@ -5,11 +5,12 @@ import numpy as np
 import pandas as pd
 from mock import MagicMock
 
+from glue.app.qt import GlueApplication
+
 from .. import qglue
 from ..core import Data
 from ..core.exceptions import IncompatibleAttribute
 from ..core.registry import Registry
-from ..qt.glue_application import GlueApplication
 from .helpers import requires_astropy
